implements slapcrafting as a component
implements slapcrafting as a component so that we can eventually blow up fragile and difficult to modify and maintain if stacks for slapcrafting recipes.

in the process of doing this ive had to overhaul a lot of other secondary things to bring them closer to contemporary tg code standards. most of this work is done, and this build compiles, but it produces a significant amount of runtimes based around about four broken things right now.