Stat Roll Emotes
Adds emotes that use stats for RP rolls. For example, by using the *strength, *str, or Roll Strength emote you can do a Strength roll based on your Strength stat, with fun success and failure messages.

This is a fun flavor and roleplay feature. It is completely optional if you want to use it for your roleplay or not.